requirements.txt
blinker==1.4
click==7.0
colorama==0.3.7
dnspython==1.16.0
Flask==1.1.4
getch==1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.4
python-dotenv==0.10.4
tqdm==4.32.2
Werkzeug==0.16.1